instruments/requirements.txt
pymeasure
pyvisa_py
pyserial
python-vxi11
pandas
